% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello1.R
\name{statwiseprogress}
\alias{statwiseprogress}
\title{Predicts Stock Price Movement for Given Stock Symbol}
\usage{
statwiseprogress(symbol)
}
\arguments{
\item{symbol}{}
}
\description{
This package predicts whether the stock price at tommorow's market close would be higher or lower compared to today's closing place.
}
\examples{
 statwiseprogress(10)

}
